add license to project
This adds the MIT copyright license to the project to make it easier
for other people to re-use. The other clones of the original zxcvbn
also seem to be doing the same thing.

This is the only identifying information I have on the author, not sure
if that's enough for you?:

    tsyrogit https://github.com/tsyrogit