makefile: add install rules
There was an existing, out of date patch in rpm/ which cannot have been
tested in a long time as it no longer even applied. This is the
inadvisable peril of applying part of a project's source code as patch
files shipped with the source.

The patch file had three additional problems:
- it used a bashism, and failed on systems where /bin/sh is the Dash
  shell
- it installed every possible *.h file, even though only zxcvbn.h is
  part of the library interface
- every single directory directive was inflexible, failed for installing
  out of /usr, and LIBDIR required passing an explicit value as the
  default was a GNU Make typo.

Rectify all these mistakes.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>